style(Files.mmOption): rename to option
