+++ src/mmHook.h
@@ -16,7 +16,7 @@
  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  ********************************************************/
 #pragma once
-#include "mmOption.h"
+#include "option.h"
 
 class CommitCallbackHook : public wxSQLite3Hook
 {
