+++ src/mmOptionMiscSettings.cpp
@@ -17,7 +17,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************/
 
 #include "mmOptionMiscSettings.h"
-#include "mmOption.h"
+#include "option.h"
 
 #include "model/Model_Checking.h"
