+++ src/model/Model_Billsdeposits.cpp
@@ -17,7 +17,7 @@
  ********************************************************/
 
 #include "Model_Billsdeposits.h"
-#include "mmOption.h"
+#include "option.h"
 #include "Model_Category.h"
 #include "Model_Account.h"
 #include "Model_Payee.h"
