+++ src/reports/categexp.cpp
@@ -21,7 +21,7 @@
 #include "htmlbuilder.h"
 
 #include "htmlbuilder.h"
-#include "mmOption.h"
+#include "option.h"
 #include <algorithm>
 #include "model/Model_Category.h"
