+++ src/transdialog.cpp
@@ -32,7 +32,7 @@
 #include "validators.h"
 #include "webapp.h"
 
-#include "mmOption.h"
+#include "option.h"
 #include "model/Model_Account.h"
 #include "model/Model_Attachment.h"
 #include "model/Model_Category.h"
