+++ tests/Makefile.in
@@ -186,7 +186,6 @@ MMEX_TESTS_OBJECTS =  \
 	mmex_tests_maincurrencydialog.o \
 	mmex_tests_minimal_editor.o \
 	mmex_tests_mmCalculator.o \
-	mmex_tests_mmOption.o \
 	mmex_tests_mmOptionAttachmentSettings.o \
 	mmex_tests_mmOptionBaseSettings.o \
 	mmex_tests_mmOptionGeneralSettings.o \
@@ -202,6 +201,7 @@ MMEX_TESTS_OBJECTS =  \
 	mmex_tests_mmhomepagepanel.o \
 	mmex_tests_mmpanelbase.o \
 	mmex_tests_mmreportspanel.o \
+	mmex_tests_option.o \
 	mmex_tests_optionsdialog.o \
 	mmex_tests_paths.o \
 	mmex_tests_payeedialog.o \
@@ -677,9 +677,6 @@ mmex_tests_minimal_editor.o: $(srcdir)/../src/minimal_editor.cpp
 mmex_tests_mmCalculator.o: $(srcdir)/../src/mmCalculator.cpp
 	$(CXXC) -c -o $@ $(MMEX_TESTS_CXXFLAGS) $(srcdir)/../src/mmCalculator.cpp
 
-mmex_tests_mmOption.o: $(srcdir)/../src/mmOption.cpp
-	$(CXXC) -c -o $@ $(MMEX_TESTS_CXXFLAGS) $(srcdir)/../src/mmOption.cpp
-
 mmex_tests_mmOptionAttachmentSettings.o: $(srcdir)/../src/mmOptionAttachmentSettings.cpp
 	$(CXXC) -c -o $@ $(MMEX_TESTS_CXXFLAGS) $(srcdir)/../src/mmOptionAttachmentSettings.cpp
 
@@ -724,6 +721,9 @@ mmex_tests_mmpanelbase.o: $(srcdir)/../src/mmpanelbase.cpp
 
 mmex_tests_mmreportspanel.o: $(srcdir)/../src/mmreportspanel.cpp
 	$(CXXC) -c -o $@ $(MMEX_TESTS_CXXFLAGS) $(srcdir)/../src/mmreportspanel.cpp
+
+mmex_tests_option.o: $(srcdir)/../src/option.cpp
+	$(CXXC) -c -o $@ $(MMEX_TESTS_CXXFLAGS) $(srcdir)/../src/option.cpp
 
 mmex_tests_optionsdialog.o: $(srcdir)/../src/optionsdialog.cpp
 	$(CXXC) -c -o $@ $(MMEX_TESTS_CXXFLAGS) $(srcdir)/../src/optionsdialog.cpp
