+++ tests/test_date_range.h
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 //----------------------------------------------------------------------------
 #include <cppunit/extensions/HelperMacros.h>
 //----------------------------------------------------------------------------
-#include "mmOption.h"
+#include "option.h"
 
 class TestFrameBase;
