+++ tests/test_relocate_category.cpp
@@ -33,7 +33,7 @@ Foundation, Inc., 59 Temple Placeuite 330, Boston, MA  02111-1307  USA
 #include "model/Model_Category.h"
 #include "model/Model_Billsdeposits.h"
 #include "model/Model_Budget.h"
-#include "mmOption.h"
+#include "option.h"
 
 /*****************************************************************************
 Turn test ON or OFF in file: defined_test_selection.h
